requirements.txt
annotated-types==0.7.0
anyio==4.9.0
black==25.1.0
certifi==2025.1.31
click==8.1.8
colorama==0.4.6
Deprecated==1.2.18
distro==1.9.0
eval_type_backport==0.2.2
exceptiongroup==1.2.2
flake8==7.1.2
griffe==1.6.3
groq==0.20.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
importlib_metadata==8.6.1
iniconfig==2.1.0
isort==6.0.1
logfire-api==3.11.0
mccabe==0.7.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
opentelemetry-api==1.31.1
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.7
pluggy==1.5.0
pycodestyle==2.12.1
pydantic==2.10.6
pydantic-ai-slim==0.0.46
pydantic-graph==0.0.46
pydantic_core==2.27.2
pyflakes==3.2.0
pytest==8.3.5
python-dotenv==1.1.0
sniffio==1.3.1
tomli==2.2.1
typing-inspection==0.4.0
typing_extensions==4.13.0
wrapt==1.17.2
zipp==3.21.0
